[BED-121] Rename internal package structure
Took 4 minutes
