+++ src/main/java/net/alphalightning/bedwars/setup/map/stages/gamemap/EmeraldSpawnerConfigurationStage.java
@@ -2,9 +2,9 @@ package net.alphalightning.bedwars.setup.map.stages.gamemap;
 
 import net.alphalightning.bedwars.BedWarsPlugin;
 import net.alphalightning.bedwars.feedback.Feedback;
-import net.alphalightning.bedwars.feedback.visual.impl.BoundingBoxRenderer;
-import net.alphalightning.bedwars.feedback.visual.impl.LootspawnerRenderer;
-import net.alphalightning.bedwars.feedback.visual.impl.LootspawnerVisualization;
+import net.alphalightning.bedwars.feedback.visual.renderer.BoundingBoxRenderer;
+import net.alphalightning.bedwars.feedback.visual.renderer.LootspawnerRenderer;
+import net.alphalightning.bedwars.feedback.visual.renderer.LootspawnerVisualization;
 import net.alphalightning.bedwars.feedback.visual.manager.VisualizationManager;
 import net.alphalightning.bedwars.game.SpawnerType;
 import net.alphalightning.bedwars.setup.map.GameMapSetup;
