+++ src/main/java/net/alphalightning/bedwars/setup/map/stages/gamemap/MinBuildHeightConfigurationStage.java
@@ -3,8 +3,8 @@ package net.alphalightning.bedwars.setup.map.stages.gamemap;
 import io.papermc.paper.event.player.AsyncChatEvent;
 import net.alphalightning.bedwars.BedWarsPlugin;
 import net.alphalightning.bedwars.feedback.Feedback;
-import net.alphalightning.bedwars.feedback.visual.impl.HeightRenderer;
-import net.alphalightning.bedwars.feedback.visual.impl.HeightVisualization;
+import net.alphalightning.bedwars.feedback.visual.renderer.HeightRenderer;
+import net.alphalightning.bedwars.feedback.visual.renderer.HeightVisualization;
 import net.alphalightning.bedwars.feedback.visual.manager.VisualizationManager;
 import net.alphalightning.bedwars.setup.map.GameMapSetup;
 import net.alphalightning.bedwars.setup.map.MapSetup;
